+++ springboot-basic/src/main/java/com/example/commandlineapplication/domain/voucher/dto/request/VoucherCreateRequest.java
@@ -1,19 +1,14 @@
-package com.example.commandlineapplication.domain.voucher.dto;
+package com.example.commandlineapplication.domain.voucher.dto.request;
 
 import com.example.commandlineapplication.domain.voucher.model.VoucherType;
+import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
 @RequiredArgsConstructor
+@Getter
 public class VoucherCreateRequest {
 
   private final VoucherType voucherType;
   private final long discountAmount;
 
-  public VoucherType getVoucherType() {
-    return voucherType;
-  }
-
-  public long getDiscountAmount() {
-    return discountAmount;
-  }
 }
